go_openrpc_reflect: problem: schema doesnt have HistoryItem during ExpandSchema mutation
spec.ExpandSchema expands schemas. As it was, this method
was only able to receive the context of the schema (and
subsequently any traversed subschemas), and thus would
fails with what were missing Definition entries, because
the method didn't have a root schema context.

So we wrap the mutation methods as closures, allowing
mutation functions to keep root vs. subschema contexts,
which fixes this issue.

Signed-off-by: meows <b5c6@protonmail.com>