go_openrpc_reflect: remove whitespaces in hacky additionalProperties workaround
additionalProperties should be allowed to be booleans,
but this is not yet supported at etclabscore/json-schema-to-types,
and throws an error with implementation at go-ethereum.

These calls are just a hacky workaround for that issue atm.

This change fixes them by removing the space between the field
and value, since the json marshaling wont have a space there.

Signed-off-by: meows <b5c6@protonmail.com>